Отрефакторил детальную страницу фильма
