+++ src/entities/movie/lib/define-movie-vote-count.ts
@@ -1,0 +1,4 @@
+export const defineMovieVoteCount = (vote_count: number) => {
+    const calcCount = vote_count / 1000
+    return calcCount > 1 ? calcCount.toFixed(2) + 'K' : vote_count
+}
